requirements.txt
python-dotenv>=1.0.1
streamlit>=1.37.0
pandas>=2.2.2
google-generativeai>=0.7.2
